Moved Write/Parse Delta state functions into their own respective files.
