+++ inc/common/sizebuf.h
@@ -25,37 +25,41 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 extern "C" {
 #endif
 
-typedef struct {
-    bool        allowoverflow;
-    bool        allowunderflow;
-    bool        overflowed;     // set to true if the buffer size failed
-    byte        *data;
-    size_t      maxsize;
-    size_t      cursize;
-    size_t      readcount;
-    uint32_t    bits_buf;
-    uint32_t    bits_left;
-    const char  *tag;           // for debugging
-} sizebuf_t;
+	/**
+	*	@brief	"SizeBuffer", used as a safe memory buffer(mainly for net-code).
+	**/
+	typedef struct {
+		bool        allowoverflow;
+		bool        allowunderflow;
+		bool        overflowed;     // set to true if the buffer size failed
+		byte        *data;
+		size_t      maxsize;
+		size_t      cursize;
+		size_t      readcount;
+		uint32_t    bits_buf;
+		uint32_t    bits_left;
+		const char  *tag;           // for debugging
+	} sizebuf_t;
 
-void SZ_Init(sizebuf_t *buf, void *data, size_t size);
-void SZ_TagInit(sizebuf_t *buf, void *data, size_t size, const char *tag);
-void SZ_Clear(sizebuf_t *buf);
-void *SZ_GetSpace(sizebuf_t *buf, size_t len);
-void SZ_WriteByte(sizebuf_t *sb, int c);
-void SZ_WriteShort(sizebuf_t *sb, int c);
-void SZ_WriteLong(sizebuf_t *sb, int c);
-void SZ_WriteString(sizebuf_t *sb, const char *s);
 
-static inline void *SZ_Write(sizebuf_t *buf, const void *data, size_t len)
-{
-    return memcpy(SZ_GetSpace(buf, len), data, len);
-}
+	void SZ_Init(sizebuf_t *buf, void *data, size_t size);
+	void SZ_TagInit(sizebuf_t *buf, void *data, size_t size, const char *tag);
+	void SZ_Clear(sizebuf_t *buf);
+	void *SZ_GetSpace(sizebuf_t *buf, size_t len);
+	void SZ_WriteByte(sizebuf_t *sb, int c);
+	void SZ_WriteShort(sizebuf_t *sb, int c);
+	void SZ_WriteLong(sizebuf_t *sb, int c);
+	void SZ_WriteString(sizebuf_t *sb, const char *s);
 
-void *SZ_ReadData(sizebuf_t *buf, size_t len);
-int SZ_ReadByte(sizebuf_t *sb);
-int SZ_ReadShort(sizebuf_t *sb);
-int SZ_ReadLong(sizebuf_t *sb);
+	static inline void *SZ_Write(sizebuf_t *buf, const void *data, size_t len)
+	{
+		return memcpy(SZ_GetSpace(buf, len), data, len);
+	}
+
+	void *SZ_ReadData(sizebuf_t *buf, size_t len);
+	int SZ_ReadByte(sizebuf_t *sb);
+	int SZ_ReadShort(sizebuf_t *sb);
+	int SZ_ReadLong(sizebuf_t *sb);
 
 // WID: C++20: In case of C++ including this..
 #ifdef __cplusplus
