+++ src/common/fifo.cpp
@@ -18,7 +18,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 #include "shared/shared.h"
 #include "common/fifo.h"
-#include "common/messages.h"
+#include "common/messaging.h"
 
 size_t FIFO_Read(fifo_t *fifo, void *buffer, size_t len)
 {
