+++ src/common/net/Q2RTXPerimentalNetChan.cpp
@@ -19,7 +19,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "shared/shared.h"
 #include "common/common.h"
 #include "common/cvar.h"
-#include "common/messages.h"
+#include "common/messaging.h"
 #include "common/net/chan.h"
 #include "common/net/Q2RTXPerimentalNetChan.h"
 #include "common/net/net.h"
