+++ src/server/server.h
@@ -28,7 +28,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "common/cvar.h"
 #include "common/error.h"
 #include "common/files.h"
-#include "common/messages.h"
+#include "common/messaging.h"
 #include "common/net/net.h"
 #include "common/net/chan.h"
 #include "common/pmove.h"
